build: comprehensive warning cleanup, correctness regression tests, and cross-platform binding fixes for v0.1.2
This is the meat of the v0.1.2 stability pin — the prior commits set up
the ABI surface and picked up the pre-existing local patches, and this
one hardens the build so "clean v0.1.2 checkout builds, tests, and
downstream-binds without patches on a fresh macOS" is actually true.

Adds an opt-in -Werror build (QSIM_WERROR=ON) that is clean on macOS
arm64 in Release. Fixes 40+ sites of -Wall/-Wextra warnings across the
entire source tree plus the Metal .mm file. Introduces regression
tests that would have caught the two worst correctness bugs uncovered
during this pass before they ever shipped.

Correctness:
 - QC_2SQRT2 constant (the Tsirelson bound used by bell_test_confirms_
   quantum and every CHSH consumer) was encoded as 3.4142 (= 2+sqrt(2))
   instead of 2.8284 (= 2*sqrt(2)). Fixed in the preceding commit; the
   new tests/unit/test_constants.c test now pins every hex-encoded
   constant to its advertised value at 1e-14.
 - QC_FINE_STRUCTURE, QC_RYDBERG, QC_ELECTRON_G were also encoded to
   garbage values (0.005, ~-1e-163, ~1e-295). Fixed to CODATA values.
 - tests/unit/test_correctness_properties.c adds property-based checks
   for the simulator's core physics: norm preservation over random
   circuits, rotation-inverse idempotence, involution of H/X/Z/CNOT
   applied twice, and bit-exact |Phi+> amplitudes from
   create_bell_state_phi_plus. All pass to 1e-12 or better.
 - tests/unit/test_quantum_state.c's test_state_purity was asserting a
   pure state (4 equal amplitudes) has purity < 1, which cannot be
   true for a normalised state vector — replaced with correct
   purity-invariant checks.

MPI:
 - Added the mpi_sendrecv wrapper that distributed_gates.c:1107 has
   been calling without a declaration or implementation (both the
   HAS_MPI and stub branches). MPI builds now link.
 - CMake test targets for test_state_partition / test_collective_ops
   are now guarded on file existence so enabling -DQSIM_ENABLE_MPI=ON
   no longer breaks the generate step.

Cross-platform bindings:
 - bindings/python/moonlab/core.py now does platform-aware library
   name resolution (dylib / so / dll) and multi-directory search.
   Previously hard-coded libquantumsim.so (Linux-only).
 - bindings/python/moonlab/__init__.py no longer crashes when
   algorithms.py binds symbols that are not yet in the dylib; the
   import is wrapped in try/except and the module exposes
   _ALGO_AVAILABLE for consumers.
 - bindings/javascript/.../build-clapack-wasm.sh now works on macOS
   (BSD sed vs GNU sed, nproc vs sysctl, CLAPACK 3.2.1's ancient
   cmake_minimum_required). WASM builds now go through end-to-end on
   a clean macOS checkout.

Warnings — a partial list:
 - -Wnewline-eof: 35 files in src/ and tools/ were missing trailing
   newlines.
 - -Wmacro-redefined: M_PI / M_PI_2 / M_PI_4 / M_SQRT2 now guarded
   with #ifndef since math.h defines them on macOS and glibc.
 - -Wunused-parameter: (void) casts added in gpu_phase, gpu_cnot,
   rdrand_get_uint64, rdseed_get_uint64, entropy_getrandom.
 - -Wunused-variable / -Wunused-but-set-variable: sqrt_gamma in
   noise, kept_mask/traced_mask in state, measure_mask in measurement,
   log2_size in entropy, theta in matrix_math, precision_part in qpe,
   A_new_size/B_new_size/stride in gpu_metal.mm — all removed or
   (void)-suppressed.
 - -Wunused-function: track_allocation stubs in memory_align,
   get_sve_vector_length in simd_dispatch, extract_substate /
   insert_substate in qpe — marked __attribute__((unused)).
 - -Wsign-compare: explicit (int)state->num_qubits casts at loop
   boundaries in measurement, entanglement, noise, grover_optimizer,
   quantum_rng.
 - -Wformat: %lu / %llu / %zu mismatches with uint64_t / size_t / int
   standardised on inttypes.h PRIu64 in hardware_entropy.c,
   health_tests.c, tests/health_tests_test.c, and one site each in
   vqe.c and tsp_logistics.c.
 - -Wdeprecated-declarations: Metal fastMathEnabled → mathMode
   migration under @available(macOS 15, *); Apple CLAPACK zheev_ calls
   silenced with a localised #pragma clang diagnostic until the
   -DACCELERATE_NEW_LAPACK migration in Phase 1G.

Build system:
 - New QSIM_WERROR CMake option, default OFF. When ON, adds -Werror
   with -Wno-error=pedantic so libomp's (external) enumerator-range
   diagnostics do not break the build. Release build on macOS arm64
   is now -Werror clean.
 - moonlab_qrng_bytes + moonlab_abi_version smoke test now runs in
   the full ctest suite; install roundtrip verified with a tiny
   out-of-tree consumer that dlopens the installed dylib.

Test status on macOS arm64 before / after this series:
 - Before: 9 of 11 tests pass (bell_test and unit_quantum_state
   fail due to the two bugs fixed above).
 - After: 13 of 13 tests pass, including the new unit_constants,
   unit_correctness_properties, and abi_moonlab_export.

+++ src/algorithms/quantum_rng.c
@@ -50,7 +50,7 @@ static double internal_random(void) {
  * @brief Measure a single qubit in computational basis
  */
 static int measure_qubit(quantum_state_t* state, int qubit) {
-    if (!state || qubit < 0 || qubit >= state->num_qubits) return 0;
+    if (!state || qubit < 0 || qubit >= (int)state->num_qubits) return 0;
 
     uint64_t mask = 1ULL << qubit;
     double prob_one = 0.0;
